requirements.txt
python-telegram-bot[job-queue]==20.7
openai==1.12.0
gspread==6.0.2
google-auth==2.27.0
pytz==2024.1
